test-project/backend/requirements.txt
fastapi
uvicorn
python-dotenv==1.0.0
requests
fastapi-cors
fastapi-mcp
pymongo
httpx
pytest
pytest-cov
httpx